.NET: .NET related: Changed using runtime libc-bootstrap.dll instead common.dll
chibias 0.34.0

+++ test/test.h
@@ -18,6 +18,6 @@ void assert(long expected, long actual, char *code);
 int printf(char *fmt, ...);
 int sprintf(char *buf, char *fmt, ...);
 int strcmp(char *p, char *q);
-int memcmp(char *p, char *q, long n);
+int memcmp(char *p, char *q, size_t n);
 void exit(int code);
 int getptrsize();
